Refactored loading of albums to preserve an order mostly consistent with what's provided by the built-in image picker controller.

+++ Classes/ELCAssetTablePicker.m
@@ -62,7 +62,7 @@ static const NSInteger MAX_THUMBNAILS_PER_ROW = 4;
 
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
-    return ceil([self.assetGroup numberOfAssets] / MAX_THUMBNAILS_PER_ROW);
+    return ceil([self.assetGroup numberOfAssets] / (float) MAX_THUMBNAILS_PER_ROW);
 }
 
 - (NSArray*)assetsForIndexPath:(NSIndexPath*)indexPath {
